+++ db/migrate/20190612133825_create_appointments.rb
@@ -1,7 +1,7 @@
 class CreateAppointments < ActiveRecord::Migration[5.2]
 	def change
 		create_table :appointments, :primary_key => :appointment_id do |t|
-		t.integer     	:encounter_id, null: false
+		t.bigint     	:encounter_id, null: false
 		t.datetime    :appointment_date, null: false
 		t.boolean		:voided, null: false, default: 0
 		t.bigint        :voided_by
